refactor(framework): Move *RecordValues
