+++ framework/py/flwr/app/__init__.py
@@ -15,6 +15,15 @@
 """Public Flower App APIs."""
 
 
+from .constants import DEFAULT_TTL
+from .typing import (
+    ConfigRecordValues,
+    ConfigScalar,
+    ConfigScalarList,
+    MetricRecordValues,
+    MetricScalar,
+    MetricScalarList,
+)
 from flwr.common.context import Context
 from flwr.common.message import Message
 from flwr.common.record import (
@@ -34,12 +43,19 @@ __all__ = [
     "Array",
     "ArrayRecord",
     "ConfigRecord",
+    "ConfigRecordValues",
+    "ConfigScalar",
+    "ConfigScalarList",
     "Context",
+    "DEFAULT_TTL",
     "Error",
     "Message",
     "MessageType",
     "Metadata",
     "MetricRecord",
+    "MetricRecordValues",
+    "MetricScalar",
+    "MetricScalarList",
     "RecordDict",
     "UserConfig",
     "UserConfigValue",
