+++ framework/py/flwr/app/error.py
@@ -19,7 +19,8 @@ from __future__ import annotations
 
 from typing import cast
 
-DEFAULT_TTL = 43200  # This is 12 hours
+from .constants import DEFAULT_TTL as DEFAULT_TTL
+
 MESSAGE_INIT_ERROR_MESSAGE = (
     "Invalid arguments for Message. Expected one of the documented "
     "signatures: Message(content: RecordDict, dst_node_id: int, message_type: str,"
