+++ framework/py/flwr/app/typing.py
@@ -1,0 +1,26 @@
+# Copyright 2026 Flower Labs GmbH. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Flower App type definitions."""
+
+
+# Value types for MetricRecord
+MetricScalar = int | float
+MetricScalarList = list[int] | list[float]
+MetricRecordValues = MetricScalar | MetricScalarList
+
+# Value types for ConfigRecord
+ConfigScalar = MetricScalar | str | bytes | bool
+ConfigScalarList = MetricScalarList | list[str] | list[bytes] | list[bool]
+ConfigRecordValues = ConfigScalar | ConfigScalarList
