+++ framework/py/flwr/serverapp/strategy/result.py
@@ -18,8 +18,8 @@
 import pprint
 from dataclasses import dataclass, field
 
+from flwr.app.typing import MetricRecordValues
 from flwr.common import ArrayRecord, MetricRecord
-from flwr.common.typing import MetricRecordValues
 
 
 @dataclass
